rename: add packages and move files
