+++ src/main/java/gift/AdminController.java
@@ -1,7 +1,7 @@
 package gift;
 
-import gift.api.ProductDao;
-import gift.api.ProductDto;
+import gift.api.product.ProductDao;
+import gift.api.product.ProductDto;
 import jakarta.validation.Valid;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
