+++ src/main/java/gift/Application.java
@@ -1,4 +1,4 @@
-package gift.api;
+package gift;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
