+++ src/main/java/gift/api/GlobalExceptionHandler.java
@@ -1,5 +1,6 @@
 package gift.api;
 
+import gift.api.member.EmailAlreadyExistsException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
