+++ src/main/java/gift/api/member/MemberRequestDto.java
@@ -1,4 +1,4 @@
-package gift.api;
+package gift.api.member;
 
 public class MemberRequestDto {
